Replace the moon.
