+++ src/moon/MoonSpaces.ts
@@ -1,8 +1,8 @@
 export enum MoonSpaces {
   LUNA_TRADE_STATION = 'm01',
-  MARE_IMBRIUM = 'm02',
-  MARE_SERENITATIS = 'm07',
-  MARE_NUBIUM = 'm15',
-  MARE_NECTARIS = 'm17',
-  MOMENTUM_VIRIUM = 'm21',
+  MARE_IMBRIUM = 'm07',
+  MARE_SERENITATIS = 'm14',
+  MARE_NUBIUM = 'm24',
+  MARE_NECTARIS = 'm26',
+  MOMENTUM_VIRIUM = 'm32',
 }
